Fix build bustage of 29235
TBR=scherkus

From 1u to static_cast<size_t>(1).

Review URL: http://codereview.chromium.org/276057

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@29238 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c8ca48ea1bfeaa037fd7dc3183b9ceddec061afd

+++ webkit/glue/webmediaplayer_impl.cc
@@ -519,7 +519,7 @@ void WebMediaPlayerImpl::OnPipelineInitialize() {
   DCHECK(MessageLoop::current() == main_loop_);
   if (pipeline_->GetError() == media::PIPELINE_OK) {
     // Only keep one time range starting from 0.
-    WebKit::WebTimeRanges new_buffered(1u);
+    WebKit::WebTimeRanges new_buffered(static_cast<size_t>(1));
     new_buffered[0].start = 0.0f;
     new_buffered[0].end =
         static_cast<float>(pipeline_->GetDuration().InSecondsF());
